requirements.txt
python-telegram-bot 
anthropic 
python-json-logger
tenacity
